added an updated types for interfaces of various custom made data types
