+++ app/types/animeFetchSource.ts
@@ -1,0 +1,4 @@
+export interface AnimeFetchSource {
+    url: string;
+    status: string;
+}
